Use base handler for import email job
